requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
firecrawl-py>=0.0.30
tavily-python>=0.3.3
pytest>=7.4.3
pytest-asyncio>=0.21.1
pydantic>=2.5.0
